Move utility methods into a utils module Change the tests to use new utils module.
